pyproject.toml
[project]
name = "deg-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.16",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.7.0",
    "pylint>=3.3.7",
]
